refactor(verdict): move handler-only Input out of the emitted contract file
