+++ internal/verdict/verdict.go
@@ -33,19 +33,6 @@ const (
 	StatusMissing = "missing" // absent from the CV
 )
 
-// Input carries everything Compute needs, all derived from real data: the role's
-// open-vacancy total, the uncovered set (total + per-skill distribution), the full
-// role skill distribution (skill slug → vacancies listing it), and the CV's parsed
-// declared/body skill sets.
-type Input struct {
-	Total           int64
-	UncoveredTotal  int64
-	UncoveredSkills map[string]int64
-	RoleSkills      map[string]int64
-	Declared        []string
-	Body            []string
-}
-
 // Verdict is the coverage result. JSON is the wire contract shared with the
 // frontend (generated to TS via cmd/gen-contracts).
 type Verdict struct {
